build(container): source directory in builder stage is no longer bound
Bind mounts are read-only but cargo tries to generate a Cargo.lock file (if it does not already exits).
Settings a custom path to a lock file is not stable yet.
We could also commit the lock file to ensure that it exists to avoid writing to the source directory.